Split e2e sample commands
