+++ spring-shell-samples/src/main/java/org/springframework/shell/samples/e2e/BaseE2ECommands.java
@@ -1,0 +1,28 @@
+/*
+ * Copyright 2022 the original author or authors.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.springframework.shell.samples.e2e;
+
+/**
+ * Base class for all e2e commands.
+ *
+ * @author Janne Valkealahti
+ */
+abstract class BaseE2ECommands {
+
+	static final String GROUP = "E2E Commands";
+	static final String REG = "e2e reg";
+	static final String LEGACY_ANNO = "e2e anno ";
+}
